fix: PR review refactorings

+++ Applications/ConsumerApi/test/ConsumerApi.Tests.Performance/tools/snapshot-creator-v2/Models/RelationshipAndMessages.cs
@@ -11,5 +11,5 @@ public record RelationshipAndMessages(
     int ReceiverIdentityAddress,
     [property: Ignore, JsonIgnore] IdentityPoolType ReceiverIdentityPoolType)
 {
-    public int NumberOfSentMessages { get; set; }
+    public long NumberOfSentMessages { get; set; }
 }
